chore: rename lib/platform/wayland.h => lib/platform/Wayland.h
